refactor: Rename the ictcg project package to core
Rename the Django project package ictcg/ to core/ and update its references
(DJANGO_SETTINGS_MODULE, ROOT_URLCONF, WSGI_APPLICATION, template/static dirs,
context processor, sitemaps import, build.js outdir, coverage --source, .gitignore).
This also fixes Dockerfile_django, which already ran `gunicorn core.wsgi` against
the (previously) nonexistent core package. The ictcg-* CSS classes are unrelated
and left as-is.
