+++ .github/workflows/ci.yml
@@ -45,7 +45,7 @@ jobs:
           ./manage.py migrate
           ./manage.py makemigrations --check --dry-run base case_studies guidelines modules navigation sponsors streams
           ./manage.py check --fail-level WARNING
-          coverage run --source ictcg,base,case_studies,guidelines,modules,navigation,sponsors,streams manage.py test
+          coverage run --source core,base,case_studies,guidelines,modules,navigation,sponsors,streams manage.py test
       - uses: codecov/codecov-action@fb8b3582c8e4def4969c97caa2f19720cb33a72f # v7.0.0
         with:
           fail_ci_if_error: true
