+++ core/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for ictcg project.
+Django settings for core project.
 
 Generated by 'django-admin startproject' using Django 4.2.2.
 
@@ -103,12 +103,12 @@ MIDDLEWARE = [
     "wagtail.contrib.redirects.middleware.RedirectMiddleware",
 ]
 
-ROOT_URLCONF = "ictcg.urls"
+ROOT_URLCONF = "core.urls"
 
 TEMPLATES = [
     {
         "BACKEND": "django.template.backends.django.DjangoTemplates",
-        "DIRS": [BASE_DIR / "ictcg" / "templates"],
+        "DIRS": [BASE_DIR / "core" / "templates"],
         "APP_DIRS": True,
         "OPTIONS": {
             "context_processors": [
@@ -117,13 +117,13 @@ TEMPLATES = [
                 "django.template.context_processors.i18n",
                 "django.contrib.auth.context_processors.auth",
                 "django.contrib.messages.context_processors.messages",
-                "ictcg.context_processors.from_settings",
+                "core.context_processors.from_settings",
             ],
         },
     },
 ]
 
-WSGI_APPLICATION = "ictcg.wsgi.application"
+WSGI_APPLICATION = "core.wsgi.application"
 
 
 # Database
@@ -203,7 +203,7 @@ STORAGES = {
 }
 
 STATICFILES_DIRS = [
-    BASE_DIR / "ictcg" / "static",  # build.js output (JavaScript + CSS)
+    BASE_DIR / "core" / "static",  # build.js output (JavaScript + CSS)
     ("images", BASE_DIR / "frontend" / "images"),  # served at /static/images/
 ]
 
@@ -309,7 +309,7 @@ if "SENTRY_DSN" in os.environ:
     )
 
 # https://docs.wagtail.org/en/stable/reference/settings.html
-WAGTAIL_SITE_NAME = "ictcg"
+WAGTAIL_SITE_NAME = "core"
 
 # Base URL to use when referring to full URLs within the Wagtail admin backend -
 # e.g. in notification emails. Don't include '/admin' or a trailing slash
