+++ core/templates/includes/more-information-module.html
