+++ core/templates/includes/supporter-logos.html
