+++ core/templates/streams/quote_block.html
