+++ core/templates/streams/richtext_with_title_block.html
