+++ core/templates/streams/simple_richtext_block.html
