+++ core/templates/streams/sponsors_block.html
