+++ core/templates/streams/supports_block.html
