+++ core/templates/wagtailadmin/admin_base.html
